modules/queue/location/consumer/requirements.txt
kafka-python
SQLAlchemy==1.3.19
psycopg2-binary